requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
PyYAML==6.0.1
Jinja2>=3.0.1,<4.0.0
python-multipart>=0.0.5,<0.0.6
boto3>=1.26.137,<2.0.0
python-dotenv>=0.19.0